remove imageUtils
